+++ src/polymorphic/base.py
@@ -79,10 +79,8 @@ class PolymorphicModelBase(ModelBase):
 
         # determine the name of the primary key field and store it into the class variable
         # polymorphic_primary_key_name (it is needed by query.py)
-        for f in new_class._meta.fields:
-            if f.primary_key and type(f) is not models.OneToOneField:
-                new_class.polymorphic_primary_key_name = f.name
-                break
+        if new_class._meta.pk:
+            new_class.polymorphic_primary_key_name = new_class._meta.pk.name
 
         return new_class